Add support for rejecting stale non-transactional reads
Summary:
Add support for non-transactional reads to specify a lower
bound HLC timestamp that the engine must be at or after to satisfy the
read. There are several limitations/caveats for this current
implementation:
1. It only throws an error, rather than supporting blocking execution
until the engine applies the specified HLC (we can evaluate extending
this in the future).
2. It only supports non-transactional queries. Because MySQL only stores
the per-database applied HLC in-memory, rather than in the transactional engine, checking the lower bound HLC on reads inside of a transaction is more complex and less meaningful

There are three cases for reads:
1. Reads inside of a rw transaction
2. Reads inside of a ro transaction
3. Reads outside of any transaction

Case 1. doesn't apply to this change because those reads should only occur against a master. That leaves cases 2. and 3. The premise is that case 2. isn't very prevalent, so HLC conditional reads/RYW consistency will initially focus on case 3. for now. Case 2 is much harder to support, since support for that requires understanding details of the engine state more deeply than the current HLC implementation allows (the in-memory database applied HLC map is always going to drift somewhat from the engine, and switching the database mid-txn is an edge case as well, among others).

Reviewed By: abhinav04sharma, bhatvinay

Differential Revision: D21223088

fbshipit-source-id: 09615df23a5