update logged gtid/executed gtid on explicit rollback
Summary:
When the before_commit hook fails in raft plugin, we rollback the trx. At this
stage the gtid_executed is already updated with the owned gtid of the trx. This
creates problems after recent changes to make the applier start from the
gtid_executed position in the log file - the trx that was rolled back might have
actually been 'consensus-committed'. This trx will never make it to the engine
causing silent data drift between the leader and the follower.

This diff attempts to fix the problem by removing logged gtid and updating
owned gtid correctly when the trx is rolled back due to a  before_commit error.
 Note that this creates holes and any subsequent trxs that enter flush stage will
end up reusing the gtid - the fix for this is on the plugin side which ensures
that once a trx is aborted, no more replicates are allowed to go through

Reviewed By: anirbanr-fb

Differential Revision: D21284798

fbshipit-source-id: c1af53a247d