Fix the bug that broke MySQL MTR Tests
Summary:
The bug was introduced in the diff D20099773

It seems like `packet_end` is not pointing to the end of the query for all sql queries.

Sometimes the query string is rewritten after it is parsed which means the address being pointed by `packet end` is not really the end of the query string. This resulted in segfaults.

I am now relying on the info from `thd` object to identify pointer to the end of query string.

An example would be -

```
Prepare stmt from "select 1";
execute stmt;
```
In above example, when `execute stmt` is parsed, the query string becomes `select 1` instead of `execute stmt`.

Reviewed By: mzait

Differential Revision: D21342897

fbshipit-source-id: b6c08df3181