Don't accept connections in admin socket until server is fully initialized
Summary:
If you set create_admin_listener_thread=1 to create a separate listener thread for admin connections, the listener thread is spawn asynchronously. So there is a chance that admin connections will start accepting connections before the remaining initialization is done, in particular it'll crash when accessing g_cached_authentication_plugins which is initialized at acl_init. As a result we are seeing random crashes across 8.0 instances, especially during upgrade.

The fix is to make sure we only accept connections when server is ready (SERVER_OPERATING) - which is true for accepting "regular" connection as well.

Adding a specific test for this unfortunately is harder than expected because none of the debug_sync facility works on mysqld_main (requires a session), and adding sleep/loop is unreliable and you can't even kill mysqld since you don't know the pid in the script. It perhaps isn't worth the effort.

Differential Revision: D21519764

fbshipit-source-id: c0aec8ac63d