Use my_rocksdb namespace for rocksdb to avoid conflicts
Summary:
This uses a new Rocksdb feature to specify namespace explicitly to avoid conflicts with raft plugin (which Raft team is seeing this in their testing) or whatever plugin there is. I've picked my_rocksdb as the new namespace (as opposed to the default rocksdb).

We use namespace alias to keep the code mostly unchanged and it is kinda nice we can keep the current rocksdb namespace in the codebase.

Reviewed By: luqun

Differential Revision: D21526578

fbshipit-source-id: 69a3702e18b