Support multiple admission control queues per-shard
Summary:
This adds multiple queues within a shard so that we can separate out different workloads on a shard. This is controlled by a session variable admission_control_queue, which can be set to a number from 0-9. This can also be set via a connection/query attribute with the key as "@admission_control_queue". In order of precedence, we first check query attributes, then connection attributes, then session variable. By default, the session variable contains value 0, so that is the default queue.

Initially, I was planning to have named queues. However, it turns out that 5.6 does not support strings as session variables, so I used integers instead. This might be something we could do in 8.0 instead.

To control priority, we assign weights to each of the queues and this is controlled by a session variable admission_control_weights which is a comma separated list of weights. The nth number in the list specifies the weight of the nth queue. Not all 10 weights have to be specified, in which case they get a default weight of 1. The intuition behind weight is such that roughly (weight of queue) / (sum of all weights) of the running pool is allocated for that queue.

A new information schema table `admission_control_queue` was added to track stats on a per-queue basis.

This also fixes an existing bug where depending on the scheduling, two thread exiting admission control can signal the same thread at the front of the queue because it hasn't had the opportunity to dequeue itself yet. This means that we're leaking slots, since we rely on threads exiting admission control always signal a new thread. This is done by having the exiting thread do the dequeue instead of waiting for the owning thread to do it.

Reviewed By: sagar0, mzait

Differential Revision: D21598431

fbshipit-source-id: 76d3687fd59