Call SetName only if the transaction is doing writes
Summary: In a previous change (D19909832), we needed the transaction to know its xid before prepare time for write unprepared, so a call was added in `rocksdb_register_tx` to set the transaction name. However, we don't need this for read only transactions, so we were doing a bit of extra work. To fix this, only call SetName on the first write we make.

Reviewed By: atish2196

Differential Revision: D21666856

fbshipit-source-id: d6e1459ab79