Convert block_create_no_primary_key to global/session var
Summary:
This change converts block_create_no_primary_key system variable
from global-only to global/session. This is a breaking change primarily
for session that changes the global variable because this change is
not applied to the local variable of this session. To accommodate this
change the client app is being fixed first to try change the session var,
and if it fails, fall back to the global var (see dependent diff). As with
the global variable, the session variable also requires super permission.

Reviewed By: hermanlee

Differential Revision: D21733105

fbshipit-source-id: 7e619db1aef