Fix slave_tx_isolation
Summary:
I'm seeing replication failure with dependency replication. In the master we have transaction A with GTID=3467495955 doing an INSERT, and followed by GTID=uuid:3467495958 doing an DELETE on the same PK. With dependency replication, it'll make sure DELETE is happening AFTER INSERT because they update the same PK (conflict) and DELETE has a higher GTID, but DELETE doesn't wait for INSERT to commit - instead we rely on the ROW lock to let DELETE block on the commit of A. In addition to that, because A and B are running in parallel, DELETE must see ROWs from INSERT before it commits, which means we need slave_tx_isolation=READ-COMMIT for dependency replication to work properly.

Unfortunately, slave_tx_isolation didn't actually work: we need to set thd->tx_isolation to set slave_tx_isolation on the thread, and it is being set from `thd.variables.transaction_isolation` in THD::init. However we only update thd.variables.transaction_isolation until much later. The fix is to move it inside THD::init so that it will always be initialized properly.

Reviewed By: Pushapgl

Differential Revision: D21734828

fbshipit-source-id: 4bc9f6f10c3