Issue#1026: MyRocks may return "Can't find record" which is unexpected
MyRocks generally uses Snapshot Checking for transactional isolation:
read rows through snapshot, if we need to modify a row - get a lock
on it and check that it hasn't been modified after snapshot was taken.

The exception is unique secondary key check which uses "READ-COMMITTED":
get a lock on the secondary key value, then scan the latest committed
data to see if there are duplicate rows.

This works, except for INSERT ... ON DUPLICATE KEY UPDATE, where
the SQL layer expects to be able to read the duplicate row if it has
got HA_ERR_FOUND_DUPP_KEY error from the storage engine.  If it
doesn't see it, the user gets a "record not found" error.

Fixed this by re-using the approach that this patch:

  469134b Avoid 2nd GetForUpdate on INSERT on DUPLICATE KEY

used for PKs: If we find a duplicate row, save it and return it to the
SQL layer when it does an index lookup to read the duplicate row.

(Since we're handling secondary keys, we save the rowid of the
duplicate row. Then, on index lookup, we read the row and return it)