Remove PLUGIN_VAR_ALLOCATED from rocksdb vars
Summary:
PLUGIN_VAR_ALLOCATED is used in 3 rocksdb variables, and in all cases it appears to be not really needed.

Background: this option is used if the variable verification function allocates memory for the value outside of thread root storage (which is released automatically at the end of the statement). Without this option the memory would be leaked because the value is copied into another memory block allocated for the variable.

In all 3 variables the value is obtained with item_val_str() which allocates memory from thread root for string items. Non-string values are not allowed and checked prior to the variable validation function is called. So there's no need to make another copy of that memory, and instead let it be handled in the regular way. Also, instead of passing the stack buffer which is not used, it is passing nullptr now.

Squash with: D17940996.

Reviewed By: lth

Differential Revision: D21891247

fbshipit-source-id: 32e6c1df77e