Backport Bug#30299881: PARSER RECURSION DOESN'T CHECK FOR STACK BOUNDARIES
Summary:
https://github.com/mysql/mysql-server/commit/5e35ecaafa
  The patch has updated the server to not fail on the parser recursion.

This fix is a backport of the upstream change from 8.0.19 to fix a stack
overflow issue in the parser. The calls to has_into_clause() were being
made in the parser and a sufficiently deep parse tree will cause a
segmentation fault.

Stack trace shows

  0x0000000004de4b7d in PT_union::has_into_clause
  0x0000000004de811d in PT_query_expression::has_into_clause
  0x0000000004de4b7d in PT_union::has_into_clause
  0x0000000004de811d in PT_query_expression::has_into_clause
  0x0000000004de4b7d in PT_union::has_into_clause
  0x0000000004de811d in PT_query_expression::has_into_clause
  0x0000000004de4b7d in PT_union::has_into_clause
  0x0000000004de811d in PT_query_expression::has_into_clause

Porting notes: The original change is dependent on a number of other
changes in the parser. I extracted the fixes for removing the
has_into_clause() and is_union() check from the parser and verifying
them during the contextualize() check instead. This patch can be dropped
when we rebase to 8.0.19 and beyond. Upstream has changed the error code
for some of the syntax issues, but this patch keeps ER_PARSE_ERROR for
now.

Squash with null

Reviewed By: pradeep1288

Differential Revision: D21943918

fbshipit-source-id: 5961c99eba5