Make sure mysqlbinlog print all Rows_log_event flags
Summary: Today we only print STMT_END_F flags in RBR events but there are other important events that needs to be printed as well (in my case the one I needed to see was `RELAXED_UNIQUE_CHECKS_F`). This prints all the supported flags and also prints the flags in hex. We also now ASSERTs if there are unknown flags and print "UNKNOWN_FLAG(0xN)" in retail.

Reviewed By: luqun

Differential Revision: D21954221

fbshipit-source-id: ddcebed41ee