purge_logs hook and safe purge log support in myraft
Summary:
Adds the ability to safely purge logs in MyRaft. A new hook `purge_logs` is
added to raft_replication. The plugin gives back the filename that is 'safe' to
purge based on its state (such as what is needed for durability, peers that are
still catching up etc). `purge_logs` hook is also an opportunity for the plugin
to garbage collect its index chunks.

Few notes:
1. If the plugin fails to read the binlog file to be purged, then it returns an
error which will prevent the server from proceeding with the purge
2. If the plugin thinks that the safe index to be retained is 1, then it returns
an error which will prevent the server from proceeding with the purge. This
 could happen when raft is enabled and there are no trx replicated through raft
when `purge logs` is called. We cannot safely determine what is safe to purge in
 this case and hence this approach.

Reviewed By: anirbanr-fb

Differential Revision: D21846797

fbshipit-source-id: 882d12d59b0