Fix growing memory usage in dependency replication
Summary:
In dependency replication mode, queries are being logged to the SQL
thread's row_query string and were not being cleared up correctly. This
logging is not necessary and slows down the SQL thread process, so skip
this step as we do on 5.6 when parsing the keys by the SQL thread.

Squash with D19036547

Reviewed By: abhinav04sharma

Differential Revision: D22005404

fbshipit-source-id: 1f0d89ea8cf