Zero init set_var::save_result
Summary: The change to log plugin vars exposed an issue with some variables that don't actually store the value but simply act as a command with a parameter. In those cases the check function doesn't store the new value in set_var::save_result, and the update function is a no-op. However the current logging code relies on getting the new value from set_var::save_result so it gets SIGFAULT when it tries to access random memory because set_var::save_result is not initialized. The simple fix is to zero initialize set_var::save_result so the pointers are nullptr and logging works correctly. However the input value is still not logged, so it could be added to the log output as a separate change.

Reviewed By: lth

Differential Revision: D22084814

fbshipit-source-id: 03cd988fded