Add tmp table and filesort disk usage/peak session tracking
Summary:
This change adds the following status variables:
- tmp_table_disk_usage
- tmp_table_disk_usage_peak
- filesort_disk_usage
- filesort_disk_usage_peak

The usage vars show the current disk usage, and the peak shows the peak usage for this connection, or since `FLUSH STATUS` is called. These session status vars do not aggregate to global status, and `FLUSH STATUS` has been fixed to only reset the peak vars, but keep the usage vars.

Tmp table usage accounts for both explicit and implicit temp tables, and filesort accounts for filesorts during query execution. So the session filesort_disk_usage will most likely be always 0. The peak can be checked after the query is done to see how much disk space the query needed.

Disk usage tracking is added through callbacks for `IO_CACHE` (filesort) and `MyISAM` (tmp tables) which report events of file size changes. The callbacks can also perform checks of disk usage against limits. This works fine for filesort because it is always destroyed at the end of the statement. However for tmp tables the size errors seem to be swallowed and often replaced with error that table is crashed and needs to be repaired. So for now the `MyISAM` callback always returns success, and the size is checked at the handler level before the insert/update operation is performed so it's safe to return an error at that point.

Tmp table usage tracking has additional twist on the secondaries where multiple workers could share tmp tables of a single session and pass them around. The replication code is fixed to correct accounting of threads when tmp table is transferred from one to another. Since in this case the status var of one thread is updated from another thread it has to be done in atomic way.

Reviewed By: sagar0, mzait

Differential Revision: D22147885

fbshipit-source-id: 21cf7a4626b