Implement sql_stats_snapshot
Summary:
This change implements atomic SQL stats snapshot functionality. It is done through a new session only var `sql_stats_snapshot`. By default, the variable is OFF so any SQL stats query will read live stats from the 3 global maps (sql_stats, sql_text and client_attrs).

When `sql_stats_snapshot` is changed to ON, the global stats mutex is acquired, the global maps are moved into the global snapshot, and replaced with empty maps, after which the mutex is released. This guarantees that snapshot is consistent and the operation is done in constant time without any copies. After this all SQL stats queries from this session will read from the global snapshot without blocking the live stats. If another session sets its own `sql_stats_snapshot` to ON, it will also become the user of now existing global snapshot. The snapshot will cease to exist when the last snapshot user (session) is gone (`sql_stats_snapshot` is set to OFF). When this happens, by default, the snapshot is merged back into the live stats (duplicate stats entries are aggregated). Optionally the snapshot can also be dropped (see below) which allows to perform atomic 'snapshot and flush' without data loss.

The status variable `sql_stats_snapshot` shows the state of the global snapshot so it's easy to check if any session is currently using it or not.

Other sessions with `sql_stats_snapshot` OFF always see live stats merged with the global snapshot on the fly if the snapshot exists. The merge is a full join by going first through the snapshot map and looking up every entry in the live stats map. If such entry is not found it is returned as is, and if found the stats are aggregated for both entries in a temp storage. Once the pass through the snapshot map is done, it also passes through the live stats map and excludes all entries that are also present in the snapshot map by doing lookups. This process can be optimized with a bit more memory used per entry, but this merge on the fly is going to be a rare operation, so I think optimizing for reduced memory usage is more important.

The behavior of `FLUSH SQL_STATISTICS` now depends on whether `sql_stats_snapshot` is ON or OFF for that session. If it's ON then only the snapshot is flushed by marking it to be dropped when the last session stops using it, instead of merging it with the live stats. If `sql_stats_snapshot` is OFF then the live stats are flushed and the snapshot is marked to be dropped. In any case, once the snapshot is marked to be dropped, the sessions with `sql_stats_snapshot` set to OFF won't see the snapshot and will only see the live stats (until the snapshot is dropped and a new one is created).

The `max_sql_stats_count` and `max_sql_stats_size` variables work in the following way with the snapshot:
- when snapshot is created, it's count and size is included in the global counters
- entries added to live stats while snapshot exists always increment the count/size counters even if a corresponding entry already exists in the snapshot (this could be fixed by acquiring the read lock on the snapshot and performing a lookup)
- when snapshot is merged back with live stats, the count/size of the duplicate entries is subtracted from the global counters so there's no double accounting any more
- if snapshot is dropped with `FLUSH SQL_STATISTICS` the snapshot count/size is subtracted immediately from the global counters.

This change also fixes a small bug with the size of `sql_findings`. It used to be added to the sql_stats size counter but was never subtracted if the findings were flushed. It is now separated from the sql_stats size and reset on flush (but still accounted against `max_sql_stats_size`).

Reviewed By: sagar0

Differential Revision: D22366526

fbshipit-source-id: c2057f99a47