Allow setting mts_dependency_replication=None in 8.0 to turn off DP for 5.6 compat
Summary:
When turning off DP people usually write mts_dependency_replication=None in config, but that's not supported in 8.0. In 8.0 mts_dependency_replication only has TBL and STMT and the on/off is independently controlled by slave_parallel_type, which isn't present in 5.6. As a result it is easy to write such config and break 8.0, as I've discovered in one prod database where 8.0 instances simply can't be brought up, and vice versa.

To make it easier to avoid such mistakes in the future, this change supports mts_dependency_replication=None to indicating falling back to database parallel replication when slave_parallel_type=DEPENDENCY. Slave_preserve_commit_order is also ignored in this case (otherwise it would error out in our checks as it is incompatible with database parallel replication). This is somewhat hacky and confusing, but in practice it aligns reasonably well with our internal usage of dependency replication. Care is taken to hide these previously global variables in local cc files and instead expose them in functions, so new usage of those variables would fail to compile (instead of silently introducing inconsistencies).

A new test is added to exercise the path when setting all dependency replication variables with mts_dependency_replication=None.

Reviewed By: abhinav04sharma

Differential Revision: D22439164

fbshipit-source-id: f891cef2933