Better MyRocks init failure cleanup
Summary:
I'm seeing crashes in prod (merlindb only so far) where rocksdb is failing to initialize due to some WritePrepared conflict (to be investigated further later), and at process shutdown we crash when calling my_rocksdb::BlockBasedTableOptions::~BlockBasedTableOptions at __run_exit_handlers, and jemalloc is in a bad state. This is because when a plugin fail to initialize, MySQL doesn't call uninitialize function, which probably makes sense as the uninitialize would crash, unprepared to clean up a partial initialized state. I've also see similar issues in the past when we are cleaning up DBOptions at process shutdown leads to a similar crash. In general it is a bad idea to depend on clean up of global variable cleanup  in destructors because the orders aren't guaranteed.

Unfortunately it's hard to fix this once and for all - ideally the best approach is to just call rocksdb_done_func to cleanup everything but it's hard to guarantee that cleanup works correctly in partial states, especially for the thread objects and the various manager code we have. Even just deleting rdb by itself leads to a crash in some cases as it asserts it holds to the only reference to some ColumnFamily objects that are held by something else.

For now I'm just going the safe route - just clean up the rdb_*_options / stats variables that we are currently seeing issues and are safe to clean up.

Differential Revision: D22551000

fbshipit-source-id: a34d658542e