Convert dependents in Log_event_wrapper to weak_ptr to avoid long chain during obj destruction
Summary:
When we destruct Log_event_wrapper obj we also end up clearing the dependents
list which was a list of shared_ptr. This sometimes caused a long chain of event
destruction which could cause stack overflow.  Converted to list of weak_ptr
instead to avoid chaining. Converting to weak_ptr is okay because dependents
cannot destruct before their dependencies are fully satisfied.

Reviewed By: hermanlee

Differential Revision: D22484662

fbshipit-source-id: 3ea94ecb9b4