log file switching as part of promotions
Summary:
When there is a promotion/demotion in the ring, a mysql instance needs to switch
its log file from binlog to apply-log (on switching to a primary role) or
apply-log to binlog (on switching to a secondary role). This diff brings in the
changes from 5.6 into 8.0. binlog management and interfaces have changed
significantly between 5.6 and 8.0 and hence this is more than a direct port.

Some notable changes compared to 5.6 implementation:
1. Binlog_ofile is a pimpl class within MYSQL_BIN_LOG which manages the stream
interface for a binlog file (as opposed to MYSQL_LOG class in 5.6). New methods
are added here to open an existing file stream and to seek to a particular
offset within a stream.
2. io_cache_type is not needed in funtions that open binlog. This is mainly because
MYSQL_BIN_LOG always opens the underlying io-cache as WRITE_CACHE (unlike
5.6 which opens relay logs in SEQ_READ_APPEND).

Differential Revision: D22428772

fbshipit-source-id: c12ab245670