ability to stop and start threads from listener queue
Summary:
Replication threads need to be stopped and/or started at certain point when the
raft ring goes through reconfiguration (for ex: new leader election).
This diff brings in the ability for the same from 5.6. It is almost a direct
port of feature in 5.6 except for some changes around channel_map handling (due
to multi-source replication in 8.0)
1. start_slave(..) and stop_slave(..) functions need additional args based on
channel_map. These are grabbed correctly from the default channel
2. A dummy LEX_MASTER_INFO and LEX_SLAVE_CONNECTION objects are passed in
(similar to what rpl_channel_service_interface is doing in 8.0)

Reviewed By: anirbanr-fb

Differential Revision: D22460948

fbshipit-source-id: 796d2e1ad86