myrocks_hotbackup for CentOS 8.0
Summary: Looks like this was missed in my earlier CentOS 8.0 work. myrocks_hotbackup is still referring to python which needs to be changed to python2, and it still uses mysqldb which isn't available in centos 8.0 and can be replaced with pymysql. There shouldn't be any perf concern here for using pymysql as the work done through SQL connection is fairly lightweight.

Reviewed By: yoshinorim

Differential Revision: D22627954

fbshipit-source-id: 62a7b5591e4