Exit admission control during queries
Summary:
Currently, a query holds an admission control slot for the
entire duration of the query. This causes a few problems:

1. This can lead to deadlocks. A transaction holding an admission control slot could be waiting on a row lock held by a transaction waiting for admission control, leading to a deadlock cycle. A possible solution is to exit admission control before waiting on a lock.
2. Admission control is mostly used to manage cpu usage. When a query is blocked for whatever reason, eg. waiting on row lock or writing to network, it is not consuming any cpu so it is artificially constraining the amount of resources a workload can consume. Generally speaking, we can allow other queries to run if a query is blocked.
3. Long running queries are generally bad queries. The current admission control system rewards them as they can hold onto their admission control slots until they complete. Instead we should preempt them to let better queries have a chance of running. Eventually, we should also deprioritize them.

To solve this, queries will now exit and re-enter admission control
during a query.

A query will exit admission control if it is blocked, or after it has
done a certain amount of work (tracked by `THD::check_yield`).
Currently, work is counted only during fetching of rows from the storage
engine. Eventually, this could also cover many more cases.

When a query tries to re-enter admission control, it will either be
queued at the front or at the back depending on why the query released
its slot. If it was due to being blocked, then it is queued at the front
on the assumption that it should finish its work quickly. This reduces
the risk of hitting query timeout and wasting cpu by throwing away work,
and if the query was holding onto locks, then it would be able to
release them more quickly, allowing other queries to run.

If the query released its slot because it has done too much work, then
it queues itself at the back, to let other queries get a chance to run.
It's possible that a long running query is holding onto row locks as
well, but this is currently unhandled. If these are due to large
transactions, then this could be handled differently in order to prevent
replication lag.

Two new global variables are added to control this:
`admission_control_wait_events` controls which events the query will exit admission control for.
`admission_control_yield_freq` controls how many units of work a query will do until a query yields to other queries.

Reviewed By: george-reynya

Differential Revision: D22708918

fbshipit-source-id: 0b25acfe896