Fixed the crash when the log_vmessage arguments get truncated while logging
Summary:
Fixed the crash when the log_vmessage arguments get truncated while logging.

vsnprintf returns number of characters that would have been written if n had been sufficiently large, not counting the terminating null character. So modified the msg_len if that is greater max buf size -1.

Porting notes: this looks to be fixed in 8.0.22

Squash with null

Differential Revision: D22741477

fbshipit-source-id: 1a8522857fa